AppPath: Rename macOS ObjC module
