+++ lib/AppPath/app_path.cmake
@@ -8,8 +8,8 @@ list(APPEND APPPATH_SRCS
 if(APPLE)
     list(APPEND APPPATH_SRCS
         ${CMAKE_CURRENT_LIST_DIR}/private/app_path_old.cpp
-        ${CMAKE_CURRENT_LIST_DIR}/private/app_path_macosx.h
-        ${CMAKE_CURRENT_LIST_DIR}/private/app_path_macosx.m
+        ${CMAKE_CURRENT_LIST_DIR}/private/app_path_macos_dirs.h
+        ${CMAKE_CURRENT_LIST_DIR}/private/app_path_macos_dirs.m
     )
 elseif(UNIX OR HAIKU)
     list(APPEND APPPATH_SRCS
