+++ lib/AppPath/private/app_path_macos_dirs.h
